requirements.txt
streamlit
pandas
numpy
openpyxl
pyarrow
dask[dataframe]
duckdb
ydata-profiling
sweetviz
python-dateutil
pydantic
tenacity
tiktoken
langchain
llama-index
scikit-learn
nltk
tqdm

openai>=1.0.0
python-dotenv>=1.0.0